Use explicit nullable primitive casting for the sake of autoporting
DEVSIX-896

+++ layout/src/main/java/com/itextpdf/layout/renderer/AbstractRenderer.java
@@ -445,8 +445,8 @@ public abstract class AbstractRenderer implements IRenderer {
         if (opacity != null && opacity < 1f) {
             PdfExtGState extGState = new PdfExtGState();
             extGState
-                    .setStrokeOpacity(opacity)
-                    .setFillOpacity(opacity);
+                    .setStrokeOpacity((float) opacity)
+                    .setFillOpacity((float) opacity);
             drawContext.getCanvas()
                     .saveState()
                     .setExtGState(extGState);
